backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=1.4.23
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
requests>=2.26.0
PyJWT>=2.8.0 